refactor(analysis): mechanical split of analysis/visualizations.py into viz_* siblings
MAJ-04 module 1/6: analysis/visualizations.py 2412 -> 58 lines (facade).

- viz_schema.py: schema constants + payload/path helpers (leaf)
- viz_animations.py: belief/GridWorld animation builders
- viz_manifest.py: GridWorld analysis manifest writer
- viz_plots.py: per-model execution-output plots
- viz_dashboard.py: cross-framework dashboards and comparisons
- visualizations.py: facade re-exporting all 31 top-level names

Pattern: 3.3.0 execute/processor.py mechanical split (12a565b2f).
2347 moved lines verified byte-identical block-by-block against the
pre-split file; only import headers differ. Package __init__, all
consumers, and every import path unchanged.

Verified: ruff 0 violations; mypy src/gnn/analysis clean; import and
facade-contract probes pass (31 names); targeted pytest
tests/analysis/test_analysis_overall.py +
tests/pipeline/test_pomdp_gridworld_cross_framework.py: 49 passed.
